Fix too small loop variables
These loop variables had narrower type than the iteration's upper bound.

+++ src/lib/VSDCharacterList.cpp
@@ -171,7 +171,7 @@ void libvisio::VSDCharacterList::handle(VSDCollector *collector) const
     return;
   if (!m_elementsOrder.empty())
   {
-    for (unsigned i = 0; i < m_elementsOrder.size(); i++)
+    for (size_t i = 0; i < m_elementsOrder.size(); i++)
     {
       auto iter = m_elements.find(m_elementsOrder[i]);
       if (iter != m_elements.end() && (0 == i || iter->second->getCharCount()))
